+++ CHANGELOG.md
@@ -2,7 +2,36 @@
 
 > Translated by ChatGPT from Chinese.
 
-## 2.2.0 (2025-07-07)
+## 2.2.3 (07/13/2025)
+
+### Optimization
+
+- Automatically hide the pinned icon when `Show Most Visited` is turned off
+
+### Fixes
+
+- Fixed the issue where previously displayed Most Visited items were not removed after turning off `Show Most Visited`
+- Fixed an issue where wallpapers would change on other clients due to cloud sync
+  > Since local images are limited by storage size and online images are restricted by
+  > permissions, wallpapers cannot be synced, so wallpaper-related data will not be synced
+
+## 2.2.2 (07/10/2025)
+
+### Optimization
+
+- Adjusted background image loading timing to minimize visibility of the loading process
+
+### Fixes
+
+- Fixed a flickering issue that occurred during each load when the background mask was enabled
+
+## 2.2.1 (07/07/2025)
+
+### Fixes
+
+- Fixed an issue where the local background selection box style was not applied
+
+## 2.2.0 (07/07/2025)
 
 ### Features
 
@@ -25,7 +54,7 @@
 - Fixed the issue where Element components always used zhCn for i18n.
 - Prevented the description of the YiYan switch from being too long in English environments.
 
-## 2.1.0
+## 2.1.0 (06/07/2025)
 
 ### Optimization
 
@@ -50,7 +79,7 @@
 
 - Refactored code to improve readability.
 
-## 2.0.0
+## 2.0.0 (03/20/2025)
 
 **Important**: Major version upgrade, please carefully read the following changelog.
 
@@ -86,7 +115,7 @@
 - Fixed an issue where clicks on some areas of the quick launch submenu did not work.
 - Fix auto focus search box not working (unfortunately Chrome prioritizes address bar).
 
-## 1.8.1
+## 1.8.1 (03/05/2025)
 
 ### Fixes
 
@@ -94,14 +123,14 @@
   when JS is not loaded (#19)
 - Fixed the logic and animation issues of "Dark mode" and "Follow system" switch</li>
 
-## 1.8.0
+## 1.8.0 (01/05/2025)
 
 ### Features
 
 - You can now use online image APIs as wallpaper.
 - Added a toggle for dark clock font.
 
-## 1.7.0
+## 1.7.0 (12/17/2024)
 
 ### Features
 
@@ -124,7 +153,7 @@
 
 - Fixed the issue where wallpaper preview did not display after switching wallpaper types multiple times in the settings page.
 
-## 1.6.1
+## 1.6.1 (12/12/2024)
 
 ### Features
 
@@ -135,7 +164,7 @@
 - Fixed the issue where the Quick Start area would flicker when loading the new tab page.
 - Optimized the style of the "Add New Shortcut" button by reducing its opacity when the mouse is not hovering over it.
 
-## 1.6.0
+## 1.6.0 (10/09/2024)
 
 ### Features
 
@@ -151,7 +180,7 @@
 - Reduced CPU usage
 - Moved the settings initialization phase earlier to avoid seeing the default settings state when opening a new tab.
 
-## 1.5.4
+## 1.5.4 (09/18/2024)
 
 ### Features
 
@@ -169,7 +198,7 @@
 - Fixed the issue where the background blur was set to a fixed value when focusing the search bar.
 - Fixed the issue of browser freezing, high CPU usage, and memory leaks when using local images as backgrounds.
 
-## 1.5.3
+## 1.5.3 (09/07/2024)
 
 ### Features
 
@@ -188,3 +217,5 @@
 - Hid Chrome-specific hints in Firefox browser.
 - Fixed the issue where search suggestion API candidates and icons could not be displayed.
 - Remove Herobrine
+
+> For earlier update logs, please refer to the commit history